# Optimise replaceAllTypeSynonyms